Remove autogenerated files
Having these in the repository make it more difficult to see differences
in the actual code, when comparing against the upstream changes, or when
modyfing the build system and regenerating them with newer autotools.
These should not be modifyed anyway as they are autogenerated, and
having them present might confuse people into doing so. They are also
quite big.

Signed-off-by: Guillem Jover <gjover@sipwise.com>