Use OpenSSL digest implementation only if GnuTLS is not available
The code checking for whether to use the OpenSSL MD5 implementation
are executed depending on GnuTLS being not available, based on the
GNUTLS_LIBS variable being empty which should be true in principle,
but seems like the wrong place to key on this. Instead move both
MD5 and SHA1 checks on the ACTION-IF-NOT-FOUND block of the GnuTLS
check.

Signed-off-by: Guillem Jover <gjover@sipwise.com>