Use automake subdir-objects option
Newer automake complain whenever there are subdirectories compiling code
from other directories. With this option we make sure the objects are
stored on the current build directory and do not interfere with the ones
they are coming from.

Use relative paths in _SOURCES, because automake does not cope well with
variables there.

Signed-off-by: Guillem Jover <gjover@sipwise.com>