x-option for constant monitoring
introduce new monitoring mode that keeps sipsak running until
failureThis new -X mode keeps sending messages until a failure occurs. A
negative answer is not considered a failure. This is helpful when you
want to monitor a SIP device and do not want to restart sipsak all over
again — it keeps running until the device fails. In this mode, the
RFC3261 retransmission timers are ignored — they are constant.

Use it like this: keep trying every 1/4 sec and return with a failure
if 5 retries are not sufficient:
sipsak -s sip:foo@bar.com -X -Z 250 -D 5

I know it is not a best practice but I did several unnecessary changes
too — the code was simply too widely nested and I had a too narrow
monitor — I hope that’s fine. If not I can submit just the minimum
necessary changes instead.